+++ pkg/collector/corechecks/windows_event_log/check_test.go
