+++ pkg/collector/corechecks/windows_event_log/unsupported_platforms.go
